Added most basic pure virtual test to HDO6 + Moved all HDO6 logic into MAUI base class
